backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
google-auth
ollama
bcrypt
PyJWT
pydantic
pytest
